chore: moving defaults to user folder
