+++ app/modules/user/validations/last_name.ts
@@ -1,7 +1,7 @@
 import getBodyField from '@/utils/express-validator/getBodyField'
 import UserRepository from '../repository/UserRepository'
 
-import { FIELD_VAR, LENGTH_VAR } from '../repository/defaults'
+import { FIELD_VAR, LENGTH_VAR } from '../defaults'
 
 const LAST_NAME_LENGTH_ERROR_MESSAGE = UserRepository.LENGTH_ERROR_MESSAGE
   .replace(FIELD_VAR, 'first_name')
