+++ app/modules/user/validations/username.ts
@@ -2,7 +2,7 @@ import { getCustomRepository } from 'typeorm'
 
 import getBodyField from '@/utils/express-validator/getBodyField'
 import UserRepository from '../repository/UserRepository'
-import { FIELD_VAR, LENGTH_VAR } from '../repository/defaults'
+import { FIELD_VAR, LENGTH_VAR } from '../defaults'
 
 const NON_UNIQUE_ERROR_MESSAGE = UserRepository.NON_UNIQUE_ERROR_MESSAGE
   .replace(FIELD_VAR, 'username')
